+++ CHANGELOG.md
@@ -14,9 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Release Energy` feature to `Isekai Protagonist` progression.
 - Added `Chuunibyou Power` feature to `Edge Lord` progression.
 - Added `Alteration of body and mind` selection to `God Emperor` progression.
+- Added `God Emperor Barrier` selection to `God Emperor` progression.
+- Added `God Emperor Aura` selection to `God Emperor` progression.
 - Added `Mastermind` Archetype (Isekai Protagonist).
 - Added `Overlord` Archetype (Isekai Protagonist).
 ### Changed
+- Reworked `God Emperor` `Divine Array` feature.
 - Buffed `Friendly Aura` to also affect damage rolls.
 - Buffed `Isekai Protagonist` to be able to choose between `Divine Aura`, `Dark Aura`, and `Friendly Aura`.
 - Buffed `Isekai Protagonist` to be able to choose two `Signature Move` abilities.
@@ -35,7 +38,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed `Isekai Protagonist` `Otherworldly Stamina` obtained from level 15 -> 13.
 - Changed `Isekai Protagonist` `Quick-Footed` obtained from level 16 -> 15.
 - Changed `God Emperor` casting stat to WIS.
-- Changed `God Emperor` `Divine Array` obtained from level 3 -> 5.
 - Changed `God Emperor` `Channel Energy` obtained from level 5 -> 3.
 - Changed `God Emperor` `Armor Saint` obtained from level 4 -> 3.
 - Renamed `Energy Immunity Selection` to `Energy Immunity`.